:adhesive_bandage: Code-review changes
Move training title logic to a dedicated file.
