+++ src/index.js
@@ -1,12 +1,5 @@
 import "./index.css";
-import slidesContent from "training-material/Slides/slides.json";
-
-const convertTextToHtml = document.createElement("div");
-convertTextToHtml.innerHTML = slidesContent[0];
-
-const trainingTitle = convertTextToHtml.querySelector(
-  "section h1:first-of-type"
-).textContent;
+import { trainingTitle } from "./title.js";
 
 document.title = trainingTitle + " 🎓";
 const pageTitle = document.querySelector(
