+++ src/labs/labs.js
@@ -1,7 +1,7 @@
 import "prismjs/themes/prism.css";
 import "./labs.css";
 import labs from "training-material/CahierExercices/parts.json";
-import slidesContent from "training-material/Slides/slides.json";
+import { trainingTitle } from "../title.js";
 
 const labsContainer = document.querySelector(".labs");
 labsContainer.innerHTML = labs.join("\n");
@@ -13,13 +13,6 @@ versionSpan.className = "version";
 versionSpan.innerHTML = MATERIAL_VERSION;
 
 coverPageContainer.appendChild(versionSpan);
-
-const convertTextToHtml = document.createElement("div");
-convertTextToHtml.innerHTML = slidesContent[0];
-
-const trainingTitle = convertTextToHtml.querySelector(
-  "section h1:first-of-type"
-).textContent;
 
 document.title = trainingTitle + " - Labs 👩‍🔬";
 const coverPageTitle = labsContainer.querySelector(
